+++ globals.d.ts
@@ -3,3 +3,21 @@
 /// <reference types="./src/maskbook/packages/polyfills/types/dom.d.ts" />
 /// <reference types="./src/maskbook/packages/polyfills/types/firefox.d.ts" />
 /// <reference types="./src/maskbook/packages/web3-telemetry/src/env.d.ts" />
+
+declare module 'dayjs-twitter' {
+    import type { PluginFunc } from 'dayjs';
+
+    declare const plugin: PluginFunc;
+    export default plugin;
+
+    declare module 'dayjs' {
+        interface Dayjs {
+            twitter(): string;
+        }
+    }
+}
+
+declare module '*.svg' {
+    const content: React.FunctionComponent<React.SVGAttributes<SVGElement>>;
+    export default content;
+}
