+++ src/components/Oembed/index.tsx
@@ -6,7 +6,7 @@ import Player from '@/components/Oembed/Player.js';
 import { fetchJSON } from '@/helpers/fetchJSON.js';
 import type { LinkDigest, OpenGraph } from '@/services/digestLink.js';
 import type { ResponseJSON } from '@/types/index.js';
-import { OpenGraphPayloadSourceType } from '@/types/openGraph.js';
+import { OpenGraphPayloadSourceType } from '@/types/og.js';
 
 import { Mirror } from './Mirror.js';
