+++ src/helpers/getPayload.ts
@@ -1,6 +1,6 @@
 import { parseJSON } from '@masknet/web3-providers/helpers';
 
-import { type MirrorPayload, OpenGraphPayloadSourceType } from '@/types/openGraph.js';
+import { type MirrorPayload, OpenGraphPayloadSourceType } from '@/types/og.js';
 
 export function getMirrorPayload(document: Document): MirrorPayload | null {
     const dataScript = document.getElementById('__NEXT_DATA__');
