--- src/types/dayjs-twitter.d.ts
@@ -1,12 +1,0 @@
-declare module 'dayjs-twitter' {
-    import type { PluginFunc } from 'dayjs';
-
-    declare const plugin: PluginFunc;
-    export default plugin;
-
-    declare module 'dayjs' {
-        interface Dayjs {
-            twitter(): string;
-        }
-    }
-}
