+++ src/types/index.ts
@@ -1,14 +1,4 @@
-import { HomeIcon } from '@heroicons/react/24/outline';
-
 import { ServerErrorCodes } from '@/helpers/createErrorResponseJSON.js';
-
-export interface NavigationItem {
-    id: string;
-    name: React.ReactNode;
-    icon: typeof HomeIcon;
-    href: string;
-    selected?: boolean;
-}
 
 export type ResponseJSON<T> =
     | {
@@ -30,8 +20,3 @@ export interface MetadataAsset {
     artist?: string;
     title?: string;
 }
-
-export type LensPublicationViewCount = {
-    id: string;
-    views: number;
-};
