--- src/types/svg.d.ts
@@ -1,4 +1,0 @@
-declare module '*.svg' {
-    const content: React.FunctionComponent<React.SVGAttributes<SVGElement>>;
-    export default content;
-}
